% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extractPlot}
\alias{extractPlot}
\alias{extractPlot.importance_plot}
\alias{extractPlot.tornado_plot}
\alias{extractData}
\alias{extractData.importance_plot}
\alias{extractData.tornado_plot}
\title{Generic for extracting a plot or data}
\usage{
extractPlot(x, ...)

\method{extractPlot}{importance_plot}(x, ...)

\method{extractPlot}{tornado_plot}(x, ...)

extractData(x, ...)

\method{extractData}{importance_plot}(x, ...)

\method{extractData}{tornado_plot}(x, ...)
}
\arguments{
\item{x}{an object of type \code{importance_plot} or \code{tornado_plot}}

\item{...}{arguments passed to other methods}
}
\value{
a plot object or data object
}
\description{
Generic for extracting a plot or data
}
