% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_glm.R
\name{importance.glm}
\alias{importance.glm}
\title{GLM variable importance plot}
\usage{
\method{importance}{glm}(model_final, model_null, dict = NA, ...)
}
\arguments{
\item{model_final}{a glm object}

\item{model_null}{a glm object for the null model}

\item{dict}{a dictionary to translate the model variables to plotting variables}

\item{...}{additional arguments}
}
\value{
a ggplot2 object
}
\description{
GLM variable importance plot
}
\examples{
gtest <- glm(mpg ~ cyl*wt*hp + gear + carb, data=mtcars, family=gaussian)
gtestreduced <- glm(mpg ~ 1, data=mtcars, family=gaussian)
g <- importance(gtest, gtestreduced)

gtest <- glm(mpg ~ cyl + wt + hp + gear + carb, data=mtcars, family=gaussian)
gtestreduced <- glm(mpg ~ 1, data=mtcars, family=gaussian)
g <- importance(gtest, gtestreduced)

gtest <- glm(vs ~ wt + disp + gear, data=mtcars, family=binomial(link="logit"))
gtestreduced <- glm(vs ~ 1, data=mtcars, family=binomial(link="logit"))
g <- importance(gtest, gtestreduced)
}
