% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_lm.R
\name{importance.lm}
\alias{importance.lm}
\title{LM variable importance plot}
\usage{
\method{importance}{lm}(model_final, model_null, dict = NA, ...)
}
\arguments{
\item{model_final}{a lm object}

\item{model_null}{a lm object for the null model}

\item{dict}{a dictionary to translate the model variables to plotting vriables}

\item{...}{additional arguments}
}
\value{
a ggplot2 object
}
\description{
LM variable importance plot
}
\examples{
gtest <- lm(mpg ~ cyl*wt*hp + gear + carb, data=mtcars)
gtestreduced <- lm(mpg ~ 1, data=mtcars)
g <- importance(gtest, gtestreduced)

gtest <- lm(mpg ~ cyl + wt + hp + gear + carb, data=mtcars)
gtestreduced <- lm(mpg ~ 1, data=mtcars)
g <- importance(gtest, gtestreduced)
}
