% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Generic Importance Plot}
\usage{
importance(model_final, ...)
}
\arguments{
\item{model_final}{a model object}

\item{...}{arguments passed to other methods}
}
\value{
an object of type importance_plot
}
\description{
Generic Importance Plot
}
