% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_importance_plot.R
\name{plot.importance_plot}
\alias{plot.importance_plot}
\title{Plot an Importance Plot object}
\usage{
\method{plot}{importance_plot}(x, ...)
}
\arguments{
\item{x}{an importance_plot object}

\item{...}{future arguments}
}
\value{
an importance plot
}
\description{
Plot an Importance Plot object
}
\examples{
gtest <- glm(mpg ~ cyl*wt*hp + gear + carb, data=mtcars, family=gaussian)
gtestreduced <- glm(mpg ~ 1, data=mtcars, family=gaussian)
g <- importance(gtest, gtestreduced)
plot(g)
}
