% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tornado_plot.R
\name{plot.tornado_plot}
\alias{plot.tornado_plot}
\title{Plot a Tornado Plot object}
\usage{
\method{plot}{tornado_plot}(x, ...)
}
\arguments{
\item{x}{a \code{tornado_plot} object}

\item{...}{future arguments}
}
\value{
the plot
}
\description{
Plot a Tornado Plot object
}
\examples{
gtest <- lm(mpg ~ cyl*wt*hp, data = mtcars)
tp <- tornado(gtest, type = "PercentChange", alpha = 0.10, xlabel = "MPG")
plot(tp)
}
