% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_importance_plot.R
\name{print.importance_plot}
\alias{print.importance_plot}
\title{print data in an \code{importance_plot}}
\usage{
\method{print}{importance_plot}(x, ...)
}
\arguments{
\item{x}{the object to be printed}

\item{...}{further arguments passed to \code{print.data.frame}}
}
\description{
print data in an \code{importance_plot}
}
\examples{
gtest <- glm(vs ~ wt + disp + gear, data=mtcars, family=binomial(link="logit"))
gtestreduced <- glm(vs ~ 1, data=mtcars, family=binomial(link="logit"))
g <- importance(gtest, gtestreduced)
print(g)
}
