% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tornado_plot.R
\name{print.tornado_plot}
\alias{print.tornado_plot}
\title{print data in a \code{tornado_plot}}
\usage{
\method{print}{tornado_plot}(x, ...)
}
\arguments{
\item{x}{the object to be printed}

\item{...}{further arguments passed to \code{print.data.frame}}
}
\description{
print data in a \code{tornado_plot}
}
\examples{
gtest <- lm(mpg ~ cyl*wt*hp, data = mtcars)
tp <- tornado(gtest, type = "PercentChange", alpha = 0.10, xlabel = "MPG")
print(tp)
}
