% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_ordered.R
\name{quantile.ordered}
\alias{quantile.ordered}
\title{Quantile for Ordered Factors}
\usage{
\method{quantile}{ordered}(x, probs = seq(0, 1, 0.25))
}
\arguments{
\item{x}{an ordered factor}

\item{probs}{the desired quatiles}
}
\value{
ordered factor levels at the desired quantiles
}
\description{
Quantile for Ordered Factors
}
\examples{
quantile(ordered(rep(c("C","B","A"), each=30), levels=c("C","B","A")),
         probs <- seq(0, 1, 0.25))
}
