% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado_glmnet.R
\name{tornado.cv.glmnet}
\alias{tornado.cv.glmnet}
\title{GLMNET Tornado Diagram}
\usage{
\method{tornado}{cv.glmnet}(
  model,
  modeldata,
  form,
  s = "lambda.1se",
  type = "PercentChange",
  alpha = 0.1,
  alt.order = NA,
  dict = NA,
  xlabel = "Response Rate",
  ...
)
}
\arguments{
\item{model}{a \code{cv.glmnet} object}

\item{modeldata}{the raw data used to fit the glmnet model}

\item{form}{the model formula}

\item{s}{Value(s) of the penalty parameter \code{lambda} at which predictions are required. Default is the value \code{s="lambda.1se"} stored on the CV object. Alternatively \code{s="lambda.min"} can be used. If s is numeric, it is taken as the value(s) of \code{lambda} to be used.}

\item{type}{PercentChange, percentiles, or ranges}

\item{alpha}{the level of change}

\item{alt.order}{an alternate order for the plot}

\item{dict}{a dictionary to translate variables for the plot}

\item{xlabel}{a label for the x-axis}

\item{...}{further arugments}
}
\value{
the plot
}
\description{
GLMNET Tornado Diagram
}
\examples{
if (requireNamespace("glmnet", quietly = TRUE))
{
  mf <- model.frame(mpg ~ cyl*wt*hp, data=mtcars)
  mm <- model.matrix(mf, mf)
  gtest <- glmnet::cv.glmnet(x = mm, y= mtcars$mpg, family = "gaussian")
  tornado(gtest, mtcars, formula(mpg ~ cyl*wt*hp), s="lambda.1se",
          type = "PercentChange", alpha = 0.10, xlabel = "MPG")
}
}
