% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado_glm.R
\name{tornado.glm}
\alias{tornado.glm}
\title{GLM Tornado Diagram}
\usage{
\method{tornado}{glm}(model, type = "PercentChange", alpha = 0.1,
  alt.order = NA, dict = NA, xlabel = "Response Rate")
}
\arguments{
\item{model}{a glm object}

\item{type}{PercentChange, percentiles, or ranges}

\item{alpha}{the level of change}

\item{alt.order}{an alternate order for the plot}

\item{dict}{a dictionary to translate variables for the plot}

\item{xlabel}{a label for the x-axis}
}
\value{
the plot invisibly
}
\description{
GLM Tornado Diagram
}
\examples{
gtest <- glm(mpg ~ cyl*wt*hp, data = mtcars, family = gaussian)
tornado(gtest, type = "PercentChange", alpha = 0.10, xlabel = "MPG")
}
