% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado_lm.R
\name{tornado.lm}
\alias{tornado.lm}
\title{Linear Model Tornado Diagram}
\usage{
\method{tornado}{lm}(
  model,
  type = "PercentChange",
  alpha = 0.1,
  alt.order = NA,
  dict = NA,
  xlabel = "Response Rate",
  ...
)
}
\arguments{
\item{model}{a lm object}

\item{type}{PercentChange, percentiles, or ranges}

\item{alpha}{the level of change}

\item{alt.order}{an alternate order for the plot}

\item{dict}{a dictionary to translate variables for the plot}

\item{xlabel}{a label for the x-axis}

\item{...}{further arguments}
}
\value{
the plot invisibly
}
\description{
Linear Model Tornado Diagram
}
\examples{
gtest <- lm(mpg ~ cyl*wt*hp, data = mtcars)
tornado(gtest, type = "PercentChange", alpha = 0.10, xlabel = "MPG")
}
