% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado.R
\name{tornado}
\alias{tornado}
\title{Generic tornado method}
\usage{
tornado(x, ...)
}
\arguments{
\item{x}{a model object}

\item{...}{other arguments passed to methods for specific classes}
}
\value{
a plot object
}
\description{
Generic tornado method
}
