% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_glm.R
\name{importance.glm}
\alias{importance.glm}
\title{GLM variable importance plot}
\usage{
\method{importance}{glm}(
  model_final,
  model_null,
  dict = NA,
  col_imp_alone = "#69BE28",
  col_imp_cumulative = "#427730",
  ...
)
}
\arguments{
\item{model_final}{a model object}

\item{model_null}{a glm object for the null model}

\item{dict}{a dictionary to translate the model variables to plotting variables}

\item{col_imp_alone}{the color used for the deviance explained by each variable
alone}

\item{col_imp_cumulative}{the color used for the cumulative deviance explained}

\item{...}{arguments passed to other methods}
}
\value{
an object of type \code{importance_plot}
}
\description{
GLM variable importance plot
}
\examples{
gtest <- glm(mpg ~ cyl*wt*hp + gear + carb, data=mtcars, family=gaussian)
gtestreduced <- glm(mpg ~ 1, data=mtcars, family=gaussian)
g <- importance(gtest, gtestreduced)

gtest <- glm(mpg ~ cyl + wt + hp + gear + carb, data=mtcars, family=gaussian)
gtestreduced <- glm(mpg ~ 1, data=mtcars, family=gaussian)
g <- importance(gtest, gtestreduced)

gtest <- glm(vs ~ wt + disp + gear, data=mtcars, family=binomial(link="logit"))
gtestreduced <- glm(vs ~ 1, data=mtcars, family=binomial(link="logit"))
g <- importance(gtest, gtestreduced)
}
