% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Generic Importance Plot}
\usage{
importance(model_final, ...)
}
\arguments{
\item{model_final}{a model object}

\item{...}{arguments passed to other methods}
}
\value{
an object of type \code{importance_plot}
\item{type}{the type of importance plot}
\item{data}{the importance data required for the plot}
}
\description{
Generic Importance Plot
}
\seealso{
\code{\link{importance.glm}} \code{\link{importance.lm}} \code{\link{importance.cv.glmnet}} \code{\link{importance.survreg}}
}
