% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_importance_plot.R
\name{plot.importance_plot}
\alias{plot.importance_plot}
\title{Plot an Importance Plot object}
\usage{
\method{plot}{importance_plot}(x, ...)
}
\arguments{
\item{x}{a \code{importance_plot} object}

\item{...}{future arguments}
}
\value{
the plot
}
\description{
Plot an Importance Plot object
}
\examples{
gtest <- lm(mpg ~ cyl + wt + hp + gear + carb, data = mtcars)
gtestreduced <- lm(mpg ~ 1, data = mtcars)
imp <- importance(gtest, gtestreduced)
plot(imp)
}
