% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado.R
\name{tornado}
\alias{tornado}
\title{Generic tornado method}
\usage{
tornado(model, ...)
}
\arguments{
\item{model}{a model object}

\item{...}{further arguments to tornado}
}
\value{
a plot object
}
\description{
Generic tornado method
}
