% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado_survreg.R
\name{tornado.survreg}
\alias{tornado.survreg}
\title{Survreg Tornado Diagram}
\usage{
\method{tornado}{survreg}(model, modeldata, type = "PercentChange",
  alpha = 0.1, alt.order = NA, dict = NA, xlabel = "Response Rate",
  ...)
}
\arguments{
\item{model}{a survreg object}

\item{modeldata}{the data used to fit the model}

\item{type}{PercentChange, percentiles, or ranges}

\item{alpha}{the level of change}

\item{alt.order}{an alternate order for the plot}

\item{dict}{a dictionary to translate variables for the plot}

\item{xlabel}{a label for the x-axis}

\item{...}{further arguments}
}
\value{
the plot invisibly
}
\description{
Survreg Tornado Diagram
}
\examples{
require(survival)
gtest <- survreg(Surv(futime, fustat) ~ ecog.ps + rx, ovarian,
                           dist='weibull', scale=1)
tornado(gtest, ovarian, type = "PercentChange", alpha = 0.10, xlabel = "futime")
}
