% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado_train.R
\name{tornado.train}
\alias{tornado.train}
\title{Caret Tornado Diagram}
\usage{
\method{tornado}{train}(
  model,
  type = "PercentChange",
  alpha = 0.1,
  alt.order = NA,
  dict = NA,
  xlabel = "Response Rate",
  ...
)
}
\arguments{
\item{model}{a \code{caret} object}

\item{type}{PercentChange, percentiles, or ranges}

\item{alpha}{the level of change}

\item{alt.order}{an alternate order for the plot}

\item{dict}{a dictionary to translate variables for the plot}

\item{xlabel}{a label for the x-axis}

\item{...}{further arugments}
}
\value{
the plot
}
\description{
Caret Tornado Diagram
}
\examples{
if (requireNamespace("caret", quietly = TRUE) &
    requireNamespace("randomForest", quietly = TRUE))
{
  gtest <- caret::train(x = subset(mtcars, select = -mpg), y = mtcars$mpg, method = "rf")
  tornado(gtest, type = "PercentChange", alpha = 0.10, xlabel = "MPG")
}
}
